requirements.txt
uvicorn
fastapi
sqlalchemy
pydantic
python-dotenv
pymysql
